% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sempl.R
\name{sempl}
\alias{sempl}
\title{Confidence sets for Causal Discovery}
\usage{
sempl(
  Y,
  B,
  O,
  innerTol = 1e-04,
  maxInnerIter = 100,
  outerTol = 1e-05,
  maxOuterIter = 1000,
  naive = F,
  Binit = NULL,
  Oinit = NULL,
  OmegaScale = 0.9,
  aelWeight = 0
)
}
\arguments{
\item{Y}{a matrix with p columns (corresponding to variables) and each row is an observation. Should be mean zero}

\item{B}{a p x p binary matrix where B[i,j] = 1 indicates a directed edge j -> i}

\item{O}{a p x p binary matrix where O[i,j] = O[j,i] = 1 indicates the bidirected edge i <-> j}

\item{innerTol}{convergence criteria for inner optimization in calculating EL LRT at specific point}

\item{maxInnerIter}{max iterations for inner optimization in calculating EL LRT}

\item{outerTol}{convergence criteria for outer optimization searching over parameters}

\item{maxOuterIter}{max iterations for outer optimization searching over parameters}

\item{naive}{if FALSE (default) profiles out bidirected parameters. If TRUE, search over all directed and bidirected parameters}

\item{Binit}{optional initialization for directed edges}

\item{Oinit}{optional initialization for bidirected edged}

\item{OmegaScale}{parameter which governs initialization}

\item{aelWEight}{weight given to extra augmented EL point. Default is 0 (i.e., no AEL point)}
}
\value{
ci is the interval
length is the length
}
\description{
Estimate parameters of SEM corresponding to ADMG
}
