% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceFunc.R
\name{var.el}
\alias{var.el}
\title{Empirical Likelihood Methods for Linear Structural Equation Models}
\usage{
var.el(Y, B, Omega, B.hat, Omega.hat, weights)
}
\description{
Calculate assymptotic variance for non-structural zeros in B and Omega matrix. The ordering of the estimates
in the variance matrix follows standard vec(-) convention indexing over rows first then columns. The elements of B
are ordered before the elements of Omega. Duplicate elements of Omega are not included.
}
